Including proper header functions

+++ Arrays_dsa/Part-02/LECTURE-10/max-subArr.cpp
@@ -1,12 +1,14 @@
 #include <iostream>
 #include <vector>
+#include <algorithm>
+#include <climits>
 using namespace std;
 
 int main() {
     int n = 5;
     int arr[5] = {1, 2, 3, 4, 5};
 
-    int maxSum = INT8_MIN;
+    int maxSum = INT_MIN;
     
     for(int st = 0; st < n; st++){
         int sum = 0;
